td-bot/scripts/requirements.txt
beautifulsoup4==4.4.1
httplib2==0.9.2
nflgame==1.2.16
pytz==2015.6
wheel==0.24.0
python-dateutil==2.5.3
